chore(wip): use commander
